180: move SvgRenderVx
