+++ src/PixelFarm/PixelFarm.MiniAgg.Complements/PixelFarm.MiniAgg.Complements.csproj
@@ -245,6 +245,7 @@
     <Compile Include="2_SpanGenerators\SpanGenGouraud.cs" />
     <Compile Include="2_SpanGenerators\SpanGenGouraudRGBA.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="SvgRenderVx\SvgRenderVx.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\PixelFarm.DrawingCanvas\PixelFarm.DrawingCanvas.csproj">
@@ -259,6 +260,9 @@
       <Project>{4d5ce655-58fa-4523-8e46-738b941e9e47}</Project>
       <Name>PixelFarm.MiniAgg</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Reference Include="System" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
